fix(charts): line charts wouldn’t render with empty data set
The line chart wouldn’t render at all if an empty array is used on the
[data] attribute, the following error would be thrown:

TypeError: Cannot read property 'xValue' of undefined

I narrowed it down to the translation of the data, where it was assumed
that the value would be defined at all times, which isn’t the case if
data is an empty array.

This fix solves the problem.